Added functionality for checkin
Added files to add a timestamp library that we probably don't need anymore as JS comes with its own Date() library that we can use for timestamps, so that's my b. But basically, we will get a  post request on the server that we will timestamp then add to the mysql server. Parsing of data and sending to SQL is still on the todo